requirements.txt
flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.5.0
flask-cors==4.0.0
dnspython==2.4.2
PyPDF2==3.0.1
python-docx==0.8.11
bson==0.5.10
httpx==0.24.1
gunicorn==21.2.0
